Fix logging and exit when 0 files to backup
When running an incremental backup we do not need to backup nothing so
added an if statement to exit when no files need to be backed up.
Log rotation was not working. Resolved by setting to root/root and
permissions set to 666 so a user can write to the log file.